xkey_helper: Fix some conversion warnings
Weird mix of size_t and int.

+++ openvpn/openssl/xkey/xkey_helper.c
@@ -81,7 +81,7 @@ xkey_load_generic_key(OSSL_LIB_CTX *libctx, void *handle, EVP_PKEY *pubkey,
     {
         print_openssl_errors();
         msg(M_WARN, "OpenSSL error: failed to load key into ovpn.xkey provider");
-		pkey = NULL;
+        pkey = NULL;
     }
     if (ctx)
     {
@@ -148,7 +148,7 @@ xkey_encode_pkcs1(unsigned char *enc, size_t *enc_len, const char *mdname,
                         MAKE_DI(sha512), MAKE_DI(sha224), MAKE_DI(sha512_224),
                         MAKE_DI(sha512_256), {0,NULL,0}};
 
-    int out_len = 0;
+    size_t out_len = 0;
     int ret = 0;
 
     int nid = OBJ_sn2nid(mdname);
@@ -165,7 +165,7 @@ xkey_encode_pkcs1(unsigned char *enc, size_t *enc_len, const char *mdname,
 
     if (tbslen != EVP_MD_size(EVP_get_digestbyname(mdname)))
     {
-        msg(M_WARN, "Error: encode_pkcs11: invalid input length <%d>", (int)tbslen);
+        msg(M_WARN, "Error: encode_pkcs11: invalid input length <%zu>", tbslen);
         goto done;
     }
 
@@ -194,13 +194,12 @@ xkey_encode_pkcs1(unsigned char *enc, size_t *enc_len, const char *mdname,
 
     out_len = tbslen + di->sz;
 
-    if (enc && (out_len <= (int) *enc_len))
+    if (enc && (out_len <= *enc_len))
     {
         /* combine header and digest */
         memcpy(enc, di->header, di->sz);
         memcpy(enc + di->sz, tbs, tbslen);
-        dmsg(D_XKEY, "encode_pkcs1: digest length = %d encoded length = %d",
-             (int) tbslen, (int) out_len);
+        dmsg(D_XKEY, "encode_pkcs1: digest length = %zu encoded length = %zu", tbslen, out_len);
         ret = true;
     }
